Rename home component
